requirements.txt
boto3==1.40.72
dotenv==0.9.9
faiss-cpu==1.12.0
numpy==2.1.3
opensearch-py==3.0.0
optimum-rbln==0.8.1
pandas==2.2.3
pillow==11.0.0
torch==2.7.0
tqdm==4.67.1
transformers==4.51.3